+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.2.0] 2025-09-30
+
+### Added 
+
+- ETL_engineering: 
+    - `DataFrameCleaner` and `ColumnCleaner` can now handle cleaning of values to 'None'. 
+    - `DataFrameCleaner.clean()`, added the option to work on a clone of the dataframe.
+    - `DataFrameCleaner.load_clean_save()`, convenient wrapper that encapsulates the entire cleaning process into a single call.
+
 ## [9.1.0] 2025-09-29 
 
 ### Changed